Implemented file system.
Methods in Gauntlet and Project now handle adding tasks from the text
file, writing the data to a binary file, and retrieving it from that
file.

Also, implemented tests for the earlier Gauntlet methods. I don't think
I'll do tests for anything else on this project. Since the whole point
of the program is to show the records, it's pretty easy to tell when
things are working.

As an aside to myself, I should probably go back to actually handle the
errors, instead of going with Eclipse's TODO stuff, but that's not
exactly a huge priority or a learning experience.